# src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"securitySchemes": {}
	},
	"info": {
		"title": "webapp",
		"version": "1.0.0",
		"description": "repository for web applications",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/sample": {
			"get": {
				"operationId": "CheckConnection",
				"responses": {
					"200": {
						"description": "server responds with 200 OK if it is healhty."
					},
					"503": {
						"description": "server responds with 503 if it is not healhty."
					}
				},
				"tags": [
					"public"
				],
				"security": [],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000",
			"description": "Development server"
		}
	]
}